+++ lib/fox_enum.dart
@@ -1,0 +1,5 @@
+/// A library for FoxEnum, a Flutter package that provides utilities
+/// for working with enums.
+library;
+
+export 'src/utils/flutter_platforms.dart';